06_gradio_app/requirements.txt
transformers>=4.40.0
torch>=2.2.0
gradio>=4.26.0
python-dotenv>=1.0.0
Pillow>=10.0.0
